simplify Desktop
