--- experimental/examples/imageviewer/shared/src/commonMain/kotlin/example/imageviewer/utils/Coroutines.kt
@@ -1,7 +1,0 @@
-package example.imageviewer.utils
-
-import kotlinx.coroutines.CoroutineScope
-import kotlin.coroutines.CoroutineContext
-import kotlin.coroutines.EmptyCoroutineContext
-
-expect fun <T> runBlocking(context: CoroutineContext = EmptyCoroutineContext, block: suspend CoroutineScope.() -> T): T
