+++ experimental/examples/imageviewer/shared/src/desktopMain/kotlin/example/imageviewer/model/filtration/PixelFilter.kt
