--- experimental/examples/imageviewer/shared/src/desktopMain/kotlin/example/imageviewer/style/Decoration.kt
@@ -1,9 +1,0 @@
-package example.imageviewer.style
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.res.painterResource
-import java.awt.image.BufferedImage
-import javax.imageio.ImageIO
-
-@Composable
-fun icAppRounded() = painterResource("ic_imageviewer_round.png")
