--- experimental/examples/imageviewer/shared/src/desktopMain/kotlin/example/imageviewer/utils/Coroutines.kt
@@ -1,9 +1,0 @@
-package example.imageviewer.utils
-
-import kotlinx.coroutines.CoroutineScope
-import kotlin.coroutines.CoroutineContext
-
-actual fun <T> runBlocking(
-    context: CoroutineContext,
-    block: suspend CoroutineScope.() -> T
-): T = kotlinx.coroutines.runBlocking(context, block)
